requirements.txt
# ── PyTorch (install FIRST, separately) ──
# GPU:  pip install torch --index-url https://download.pytorch.org/whl/cu126
# CPU:  pip install torch

# ── RLGym + RocketSim + RLViser ──
# pip install rlgym[rl-sim,rl-rlviser]

# ── PPO training engine (from GitHub) ──
# pip install git+https://github.com/AechPro/rlgym-ppo

# ── Other dependencies ──
numpy==1.26.4
gym==0.26.2
pillow==12.1.1
cloudpickle==3.1.2
rlgym_tools==2.6.0
numba==0.64.0
scipy==1.15.3

# ── SB3 (for mechanic curriculum training) ──
stable-baselines3>=2.0
tensorboard

# Optional (for eval GIF rendering, logging)
matplotlib==3.10.8
wandb==0.25.0
